pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "let-me-sync"
version = "0.1.0"
authors = [
    {name = "Martanto", email = "martanto@live.com"},
]
description = "FastAPI server to sync data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "python-multipart",
    "jinja2",
    "aiofiles",
    "bcrypt",
    "python-dotenv",
    "itsdangerous>=2.2.0",
    "pandas>=3.0.1",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "zipstream-ng>=1.7.1",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.uv]
cache-dir = "./.cache/uv"
package = true

[tool.hatch.build.targets.wheel]
packages = ["server"]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[project.scripts]
let-me-sync = "server.__main__:main"
server-seed = "server.cli.seed:seed"
server-refresh = "server.cli.refresh:refresh"

[project.urls]
Homepage = "https://github.com/martanto/let-me-sync"
Issues = "https://github.com/martanto/let-me-sync/issues"
